src/requirements/base.txt
Django==3.1.3
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-jsonapi==3.1.0
djangorestframework-simplejwt==5.0.0

requests==2.26.0
